:sushi: Implement independent hashers properly
Pull out hashing into another trait (`IntoSketch`) that defines how to
turn an implementor into a u64. Use that as a translation layer into the
actual u64 we want to start counting (this is kind of like the contramap
stuff algebird does, I guess).

Means we can pull `SipHasher` out of the core processing (although some
standard implementations are provided for convenience/testing anyway, so
the imports are still around).

Also means we can include a multiplying random value to each row as a
"hasher". The paper says this isn't good enough, but both ezyang's ocaml
and twitters scala implementations claims it is good enough so... What
it does mean is that addition becomes a little harder to do properly, as
you need to have sketches with the same random offset
values. Interestingly means implementing a monoid instance
properly (i.e. a distinguished zero object in a type) requires types
dependent on their hashers. Decided not to try that for now...